requirements.txt
discord.py
aiosmtplib
